pyproject.toml
[tool.poetry]
name = "mastodon-rendaino-tombstone"
version = "0.1.0"
description = ""
authors = ["kelvin27315"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/kelvin27315/mastodon-rendaino-tombstone"

[tool.poetry.dependencies]
python = "3.11.6"
pandas = "^2.2.3"
mastodon-py = "^1.8.1"


[tool.poetry.group.dev.dependencies]
pysen = "^0.11.0"
black = "^23.12.1"
flake8 = "^7.2.0"
isort = "^5.13.2"
mypy = "^1.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false
mypy_preset = "strict"
line_length = 120
py_version = "py311"

[[tool.pysen.lint.mypy_targets]]
  paths = ["./src"]
